+++ packages/core/accounts/utils/toKernelPluginManager.ts
@@ -128,11 +128,10 @@ export async function toKernelPluginManager<
             const enableSignature =
                 await getPluginEnableSignature(accountAddress)
             return getEncodedPluginsDataV2({
-                accountAddress,
-                action,
                 enableSignature,
                 userOpSignature,
-                validator: regular
+                action,
+                enableData: await regular.getEnableData(accountAddress)
             })
         } else if (sudo) {
             return userOpSignature
@@ -207,7 +206,38 @@ export async function toKernelPluginManager<
         ])
     }
 
+    const getPluginsEnableTypedData = async (accountAddress: Address) => {
+        if (!action) {
+            throw new Error("Action data must be set")
+        }
+        if (!sudo)
+            throw new Error(
+                "sudo validator not set -- need it to enable the validator"
+            )
+        if (!regular) throw new Error("regular validator not set")
+
+        const { version } = await accountMetadata(
+            client,
+            accountAddress,
+            entryPointAddress
+        )
+
+        const validatorNonce = await getKernelV3Nonce(client, accountAddress)
+
+        const typedData = await getPluginsEnableTypedDataV2({
+            accountAddress,
+            chainId,
+            kernelVersion: version,
+            action,
+            validator: regular,
+            validatorNonce
+        })
+
+        return typedData
+    }
+
     return {
+        sudoValidator: sudo,
         ...activeValidator,
         getIdentifier,
         encodeModuleInstallCallData: async (accountAddress: Address) => {
@@ -333,6 +363,7 @@ export async function toKernelPluginManager<
             return encodedNonceKey
         },
         getPluginEnableSignature,
+        getPluginsEnableTypedData,
         getValidatorInitData: async () => {
             if (validatorInitData) return validatorInitData
             return {
@@ -344,6 +375,9 @@ export async function toKernelPluginManager<
 
                 identifier: pad(getIdentifier(true), { size: 21, dir: "right" })
             }
+        },
+        signUserOperationWithActiveValidator: async (userOperation) => {
+            return activeValidator.signUserOperation(userOperation)
         }
     }
 }