+++ packages/test/v0.7/ecdsaKernelAccount.test.ts
@@ -59,6 +59,7 @@ import {
     index,
     kernelVersion,
     mintToAccount,
+    validateEnvironmentVariables,
     waitForNonceUpdate
 } from "./utils.js"
 
@@ -74,17 +75,6 @@ const requiredEnvVars = [
     "ZERODEV_BUNDLER_RPC_HOST",
     "ZERODEV_PAYMASTER_RPC_HOST"
 ]
-
-const validateEnvironmentVariables = (envVars: string[]): void => {
-    const unsetEnvVars = envVars.filter((envVar) => !process.env[envVar])
-    if (unsetEnvVars.length > 0) {
-        throw new Error(
-            `The following environment variables are not set: ${unsetEnvVars.join(
-                ", "
-            )}`
-        )
-    }
-}
 
 validateEnvironmentVariables(requiredEnvVars)