+++ packages/test/v0.7/recoveryKernelAccount.test.ts
@@ -22,7 +22,8 @@ import {
     getPublicClient,
     getRecoveryKernelAccount,
     getZeroDevPaymasterClient,
-    kernelVersion
+    kernelVersion,
+    validateEnvironmentVariables
 } from "./utils.js"
 
 dotenv.config()
@@ -37,17 +38,6 @@ const requiredEnvVars = [
     "ZERODEV_BUNDLER_RPC_HOST",
     "ZERODEV_PAYMASTER_RPC_HOST"
 ]
-
-const validateEnvironmentVariables = (envVars: string[]): void => {
-    const unsetEnvVars = envVars.filter((envVar) => !process.env[envVar])
-    if (unsetEnvVars.length > 0) {
-        throw new Error(
-            `The following environment variables are not set: ${unsetEnvVars.join(
-                ", "
-            )}`
-        )
-    }
-}
 
 validateEnvironmentVariables(requiredEnvVars)