+++ plugins/permission/index.ts
@@ -1,5 +1,9 @@
 export type * from "./types.js"
 export { toPermissionValidator } from "./toPermissionValidator.js"
 export { serializePermissionAccount } from "./serializePermissionAccount.js"
+export {
+    type MultiChainPermissionAccountsParams,
+    serializeMultiChainPermissionAccounts
+} from "./serializeMultiChainPermissionAccounts.js"
 export { deserializePermissionAccount } from "./deserializePermissionAccount.js"
 export * from "./constants.js"